LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/containment-expected.txt


FAIL canvas child containment assert_equals: expected "paint" but got "none"
